+++ scripts/ci/cache/_common.mjs
@@ -1,16 +1,12 @@
 // @ts-check
-import { readFile } from 'node:fs/promises'
 import { dirname, resolve } from 'path'
 import { fileURLToPath } from 'url'
+import { ENV_HASHES_KEY } from './_const.mjs'
 
 const __dirname = dirname(fileURLToPath(import.meta.url))
 export const SCRIPTS_DIR = __dirname
 export const ROOT = resolve(__dirname, '..', '..', '..')
 export const MOBILE_APP_DIR = resolve(ROOT, ...'apps/native/app'.split('/'))
-export async function getPackageJSON(
-  root = ROOT,
-  filePath = resolve(root, 'package.json'),
-) {
-  const content = JSON.parse(await readFile(filePath, 'utf-8'))
-  return content
-}
+export const HAS_HASH_KEYS = !!(
+  process.env[ENV_HASHES_KEY] && process.env[ENV_HASHES_KEY] != 'false'
+)
