+++ scripts/ci/cache/_const.mjs
@@ -4,3 +4,5 @@ export const ENV_INIT_CACHE = '_INIT_CACHE'
 export const ENV_KEYS = '_CACHE_KEYS'
 export const ENV_YAML_FILE = 'YAML_FILE'
 export const ENV_CACHE_SUCCESS = '_CACHE_SUCCESS'
+
+export const HASH_GENERATE_FILES_SCRIPT = 'scripts/_hash-generated-files.sh'
