+++ scripts/ci/cache/_utils.mjs
@@ -1,14 +1,15 @@
 // @ts-check
 import { arch, platform } from 'node:os'
 import crypto from 'node:crypto'
-import { ROOT, getPackageJSON } from './_common.mjs'
+import { ROOT } from './_common.mjs'
 import { exec } from 'node:child_process'
 import { resolve, join } from 'node:path'
 import { readFile, readdir, stat } from 'node:fs/promises'
+import { HASH_GENERATE_FILES_SCRIPT } from './_const.mjs'
 
 const GENERATE_HASH_GENERATED_FILES_SCRIPT = resolve(
   ROOT,
-  ...'scripts/_hash-generated-files.sh'.split('/'),
+  ...HASH_GENERATE_FILES_SCRIPT.split('/'),
 )
 export async function getGeneratedFileHash(
   scriptPath = GENERATE_HASH_GENERATED_FILES_SCRIPT,
@@ -56,6 +57,12 @@ export async function getPackageHash(
   }, {})
   return crypto.createHash('sha256').update(JSON.stringify(value)).digest('hex')
 }
+/**
+ * Calculates the SHA256 hash of the content of a yarn.lock file.
+ * @param {string} [root=ROOT] - The root directory of the project.
+ * @param {string} [filePath] - The path to the yarn.lock file.
+ * @returns {Promise<string>} The SHA256 hash of the yarn.lock file content.
+ */
 export async function getYarnLockHash(
   root = ROOT,
   filePath = resolve(root, 'yarn.lock'),
@@ -64,6 +71,14 @@ export async function getYarnLockHash(
   return crypto.createHash('sha256').update(content).digest('hex')
 }
 
+/**
+ * Checks if the file size is equal to or greater than the specified size.
+ *
+ * @param {string} filePath - The path to the file.
+ * @param {number} size - The size to compare against.
+ * @returns {Promise<boolean>} - A promise that resolves to true if the file size is equal to or greater than the specified size, otherwise false.
+ * @throws {Error} - If there is an error getting the file size.
+ */
 export async function fileSizeIsEqualOrGreaterThan(filePath, size) {
   try {
     const stats = await stat(filePath)
@@ -73,6 +88,14 @@ export async function fileSizeIsEqualOrGreaterThan(filePath, size) {
   }
 }
 
+/**
+ * Calculates the total size of a folder and checks if it is equal to or greater than a specified size.
+ * @param {string} folderPath - The path to the folder.
+ * @param {number} size - The size to compare against.
+ * @param {boolean} [number=false] - Optional flag to return the total size as a number instead of a boolean.
+ * @returns {Promise<boolean|number>} - Returns true if the folder size is equal to or greater than the specified size. If the 'number' flag is set to true, it returns the total size as a number.
+ * @throws {Error} - Throws an error if there is an issue reading the folder.
+ */
 export async function folderSizeIsEqualOrGreaterThan(
   folderPath,
   size,
@@ -105,6 +128,7 @@ export async function folderSizeIsEqualOrGreaterThan(
 }
 
 /**
+ * Run command in a child process.
  * @param {string} cmd
  * @param {string | undefined} cwd
  */
@@ -122,4 +146,19 @@ export async function runCommand(cmd, cwd = undefined) {
       resolve(stdout)
     })
   })
+}
+
+/**
+ * Retrieves the content of the package.json file.
+ *
+ * @param {string} [root=ROOT] - The root directory path.
+ * @param {string} [filePath] - The path to the package.json file.
+ * @returns {Promise<Object>} - A promise that resolves to the content of the package.json file.
+ */
+export async function getPackageJSON(
+  root = ROOT,
+  filePath = resolve(root, 'package.json'),
+) {
+  const content = JSON.parse(await readFile(filePath, 'utf-8'))
+  return content
 }
